requirements.txt
bcrypt==3.1.1
boto==2.45.0
boto3==1.4.2
botocore==1.4.88
cffi==1.8.3
coverage==4.2
dj-database-url==0.4.1
Django==1.7
django-braces==1.10.0
django-postgrespool==0.3.0
django-storages==1.5.1
docutils==0.13.1
futures==3.0.5
gunicorn==19.6.0
jmespath==0.9.0
Pillow==3.4.1
psycopg2==2.6.2
pycparser==2.17
python-dateutil==2.6.0
pytz==2016.10
s3transfer==0.1.10
six==1.10.0
SQLAlchemy==1.1.4
